Actually fixed player death crash this time
InfoBar was triggering update_health which was trying to create a tween based on the scene tree when the tree wasn't loaded anymore. I fixed this by creating a static reference to the tree as it was when it was @onready and then creating the tween based on that tree. This has the effect of making the tween and attaching it to the tree when the tree is loaded and doing nothing when the tree is not loaded.